// lib/domain/main.types.ts
export type YoutubeCaptionType = 'AUTO_GENERATED' | 'USER_GENERATED' | 'NONE'
// ...
export type YoutubeLiveStreamingDetails = Readonly<{
    actualStartTime?: string
    actualEndTime?: string
    scheduledStartTime?: string
    activeLiveChatId?: string
    scheduledEndTime?: string
    concurrentViewers?: string
}>

export type YoutubeLifeBroadcastStatus = 'none' | 'upcoming' | 'live'

export type YoutubeVideoDetails = Readonly<{
    videoId: string
    publishedAt: string
    videoTitle: string
    videoDescription: string
    channelTitle: string
    channelId: string
    channelUri: string
    liveBroadcastContent: YoutubeLifeBroadcastStatus
    duration: string
    durationSeconds: number
    liveStreamingDetails?: YoutubeLiveStreamingDetails
    privacyStatus: 'public' | 'private' | 'unlisted'
    captions: YoutubeCaptionType
    isMembersOnly: boolean
}>